==> aitelier-ml/notebooks/1_articles_feature_engineering.ipynb <==
# Articles Feature Engineering

Define constants and configurations

from typing import List

import torch
from loguru import logger
from tqdm.auto import tqdm

FEATURES_EMBEDDING_MODEL_ID: str = 'all-MiniLM-L6-v2'

Unzip the datasets

import zipfile
from pathlib import Path

# Setup path to data folder
data_path = Path('../data/')
unzipped_data_path = data_path / 'unzipped'

# If the image folder does not exists, download the image data
if not unzipped_data_path.is_dir():
    print(f'{unzipped_data_path} does not exist, creating one...')
    unzipped_data_path.mkdir(parents=True, exist_ok=True)

# Unzip the raw articles data
with zipfile.ZipFile(data_path / 'raw' / 'articles.csv.zip', 'r') as zip_ref:
    print('Unzipping articles dataset...')
    zip_ref.extractall(unzipped_data_path)

import pandas as pd

articles_df = pd.read_csv(unzipped_data_path / 'articles.csv')
articles_df.shape

articles_df.head(5)

import random

# See a sample article data
articles_df.iloc[random.randint(0, len(articles_df))]

# Let's examine the colums data types
articles_df.info()

# And also check for NaN values
articles_df.isna().sum()

As we can see, the `article_id` is a number, we will convert it to a string just to make it clear that the number does not represent any numerical significance. Additionally, there are some missing values for the `detail_desc` column. To solve this issue, we will create our own product description which summarises the appearance, color, and category.

# A utility function to create an article description.
def create_article_description(row):
    description = (
        f"{row['prod_name']} - {row['product_type_name']} in {row['product_group_name']}"
        f"\nAppearance: {row['graphical_appearance_name']}"
        f"\nColor: {row['perceived_colour_value_name']} {row['perceived_colour_master_name']} ({row['colour_group_name']})"
        f"\nCategory: {row['index_group_name']} = {row['section_name']} - {row['garment_group_name']}"
    )

    # If detail_desc for the particular sample exists, append it to the current description
    if pd.notna(row['detail_desc']) and row['detail_desc']:
        description += f"\nDetails {row['detail_desc']}"

    return description

# A utility function to compute the articles features and update the dataframe
def compute_article_features(df):
    # Now, we modify the original article dataframe
    df = df.assign(
        article_id=df['article_id'].astype(str),
        article_description=df.apply(create_article_description, axis=1),
    )

    # Drop columns with any nulls
    df = df.dropna(axis=1, how='any')

    # Remove detail_desc column as its information is already included in article_description
    cols_to_drop = [
        column
        for column in ['detail_desc', 'detail_desc_length']
        if column in df.columns
    ]
    df = df.drop(columns=cols_to_drop)

    return df

articles_df = compute_article_features(articles_df)
articles_df.shape

articles_df.head(3)

Next, we will turn the article descriptions into their embedding counterpart For this purpose, we will use HuggingFace's [SentenceTransformer](https://huggingface.co/sentence-transformers), specifically we will use the "all-MiniLM-L6-v2" embedding model. (There isn't exact reason why I use this particular model. I use it because it is what shown on HF's docs)

This embedding is the one that will be fed to the recommender model (the candidate tower).

from sentence_transformers import SentenceTransformer

for i, desc in enumerate(articles_df['article_description'].head(n=3)):
    logger.info(f'Item {i+1}:\n{desc}')

device = (
    'cuda'
    if torch.cuda.is_available()
    else 'mps'
    if torch.backends.mps.is_available()
    else 'cpu'
)
logger.info("Loading '{FEATURES_EMBEDDING_MODEL_ID}' model to {device}...")

model = SentenceTransformer(FEATURES_EMBEDDING_MODEL_ID, device=device)

# Utility function to generate the embeddings and add it to the dataframe
def generate_embeddings_for_dataframe(
    df, text_column: str, model: SentenceTransformer, batch_size: int = 32
):
    total_rows = len(df)
    pbar = tqdm(total=total_rows, desc='Generating embeddings')

    texts: List[str] = df[text_column].tolist()

    text_embeddings = []
    for i in range(0, len(texts), batch_size):
        batch_texts = texts[i : i + batch_size]
        batch_embeddings = model.encode(
            batch_texts,
            show_progress_bar=False,
            device=device,
        )
        text_embeddings.extend(batch_embeddings)
        pbar.update(len(batch_texts))

    pbar.close()
    df_with_embeddings = df.copy()
    df_with_embeddings['embeddings'] = text_embeddings

    return df_with_embeddings

articles_df = generate_embeddings_for_dataframe(
    articles_df, text_column='article_description', model=model, batch_size=128
)

Now, for each article description, we have the vector embeddings which we can feed into the model.

articles_df[['article_description', 'embeddings']].head(3)

Next, we pickle it to save the processed articles data.

articles_df.to_pickle(data_path / 'processed' / 'articles.pkl')
