==> aitelier-ml/notebooks/2_customers_feature_engineering.ipynb <==
# Customers Feature Engineering

import random

import pandas as pd

import zipfile
from pathlib import Path

# Setup path to data folder
data_path = Path('../data/')
unzipped_data_path = data_path / 'unzipped'

# Unzip the raw customers data
with zipfile.ZipFile(data_path / 'raw' / 'customers.csv.zip', 'r') as zip_ref:
    print('Unzipping customers dataset...')
    zip_ref.extractall(unzipped_data_path)

### Data Inspection

First, we inspect the customers data

customers_df = pd.read_csv(unzipped_data_path / 'customers.csv')
customers_df.shape

customers_df.head(5)

customers_df.iloc[random.randint(0, len(customers_df))]

customers_df.isnull().sum()

### Feature Selection

Apparently, there are many rows with null values. We need to take care of those. There are many ways to do this, but we can start with choosing which features may be relevant in the development of our recommender system:
- customer_id
- club_member_status
- age
- postal_code

# Define the columns we want
required_columns = ['customer_id', 'club_member_status', 'age', 'postal_code']

# Check if the columns exists
missing_columns = [col for col in required_columns if col not in customers_df.columns]
if missing_columns:
    raise ValueError(f"Columns {', '.join(missing_columns)} not found in the DataFrame")

# Keep the columns we want
customers_df = customers_df[required_columns]

### Handling Missing Values in `club_members_status`

Next, let's take care of the missing values in the `club_member_status`. First, we check the possible values in this column.

print(customers_df['club_member_status'].value_counts())

print('\nNull values in club_member_status:')
print(len(customers_df) - customers_df['club_member_status'].value_counts().sum())

There are multiple possible reasons why there are null values for this column. Normally, we should consult with the data engineers or back-end engineers to clarify what does null mean in the system. But, for now, let's assume null values as not being a member. We will replace the null values with `ABSENT`.

customers_df['club_member_status'].fillna('ABSENT', inplace=True)
print(customers_df['club_member_status'].value_counts())

### Handling Missing `Age` Values and Customer Segmentation

There are many ways to handle missing values for age (e.g. imputation). However, for now, let's make it simple by simply dropping rows with null age values.

print('Null values in age before dropping:')
print(customers_df['age'].isnull().sum())

customers_df.dropna(subset=['age'], inplace=True)

print('\nNull values in age after dropping:')
print(customers_df['age'].isnull().sum())

Then, instead of using the age column as is, let's create aditional features that determines the age group of the customer. We will group the customers into the following age groups:
- 0-18
- 19-25
- 26-35
- 36-45
- 46-55
- 56-65
- 66+

def create_age_group(age):
    if age <= 18:
        return '0-18'
    elif 19 <= age <= 25:
        return '19-25'
    elif 26 <= age <= 35:
        return '26-35'
    elif 36 <= age <= 45:
        return '36-45'
    elif 46 <= age <= 55:
        return '46-55'
    elif 56 <= age <= 65:
        return '56-65'
    else:
        return '66+'


customers_df['age_group'] = customers_df['age'].apply(create_age_group)

customers_df.head(5)

customers_df.shape

### Save the Engineered Features
Lastly, we pickle it to save the processed customers data

customers_df.to_pickle(data_path / 'processed' / 'customers.pkl')
