==> aitelier-ml/notebooks/3_transactions_feature_engineering.ipynb <==
# Transactions Feature Engineering

import random

import numpy as np
import pandas as pd

import zipfile
from pathlib import Path

# Setup path to data folder
data_path = Path('../data/')
unzipped_data_path = data_path / 'unzipped'

# Unzip the raw customers data
with zipfile.ZipFile(data_path / 'raw' / 'transactions_train.csv.zip', 'r') as zip_ref:
    print('Unzipping transactions dataset...')
    zip_ref.extractall(unzipped_data_path)

### Data Inspection

First, we inspect the transactions data

transactions_df = pd.read_csv(unzipped_data_path / 'transactions_train.csv')
transactions_df.shape

transactions_df.head(5)

transactions_df.iloc[random.randint(0, len(transactions_df))]

transactions_df.isnull().sum()

transactions_df.info()

Seems like the dataset is pretty clean already. We only need to:
- Convert the `article_id` into string
- Convert `t_dat` to epoch milliseconds.
- Extract year, month, day, and day of week from `t_dat`

transactions_df = transactions_df.assign(
    article_id=lambda df: df['article_id'].astype(str),  # Convert article_id to string
    t_dat=lambda df: pd.to_datetime(df['t_dat']),  # Convert t_dat to datetime
)

# Extract year, month, day, and day of week from t_dat
transactions_df = transactions_df.assign(
    year=lambda df: df['t_dat'].dt.year,
    month=lambda df: df['t_dat'].dt.month,
    day=lambda df: df['t_dat'].dt.day,
    day_of_week=lambda df: df['t_dat'].dt.weekday,  # Monday=0, Sunday=6
)

# Calculate cyclical encodings for month
C = 2 * np.pi / 12
transactions_df = transactions_df.assign(
    month_sin=lambda df: np.sin(C * df['month']),
    month_cos=lambda df: np.cos(C * df['month']),
)

# Turn t_dat into epoch milliseconds
transactions_df = transactions_df.assign(
    t_dat=lambda df: df['t_dat'].astype(int) // 10**6
)

transactions_df.head(5)

### Save the Engineered Features
Lastly, we pickle it to save the processed transactions data

transactions_df.to_pickle(data_path / 'processed' / 'transactions.pkl')
