==> aitelier-ml/notebooks/5_ranking_data_generation.ipynb <==
# Ranking Data Generation
Interaction data provides insight into the likelihood that a customer will engage with an item (e.g., through clicks or purchases). However, this dataset is limited to customer and product attributes, meaning it does not leverage additional transaction-level features such as purchase date (which may capture seasonality), price paid, or sales channel which are factors that could influence customer behaviour. To incorporate these variables, a separate ranking model must be developed and trained using a dataset specifically designed to include these transaction characteristics.

### Load Sampled Transaction, Customer, and Articles Data

from pathlib import Path

import numpy as np
import pandas as pd

# Setup path to data folder
data_path = Path('../data/')
processed_data_path = data_path / 'processed'

transactions_df: pd.DataFrame = pd.read_pickle(
    processed_data_path / 'sampled_transactions.pkl'
)
customers_df: pd.DataFrame = pd.read_pickle(
    processed_data_path / 'sampled_customers.pkl'
)
all_articles_df: pd.DataFrame = pd.read_pickle(processed_data_path / 'articles.pkl')

# Only use article_id and customer_id from the transactions data
transactions_df = transactions_df[['customer_id', 'article_id']]

# Use all features of the articles data except article description and the embeddings.
articles_df = all_articles_df.drop(columns=['article_description', 'embeddings'])

# Only use customer_id and age from the customers data
customers_df = customers_df[['customer_id', 'age']]

# Merge the datasets
df = transactions_df.merge(articles_df, on='article_id', how='left')
df = df.merge(customers_df, on='customer_id', how='left')

df.head(5)

### Create Query Features

query_features = ['customer_id', 'age', 'article_id']
df = df[query_features]

# Create positive pairs
positive_pairs = df.copy()

Then, we create the negative pairs

RANDOM_SEED = 42
np.random.seed(RANDOM_SEED)

# Calculate the number of negative samples
n_negative = len(positive_pairs) * 10

# Sample negative article_ids, customer_ids, and other features
article_ids = np.random.choice(
    articles_df['article_id'].unique(), replace=True, size=n_negative
)
customer_ids = np.random.choice(
    customers_df['customer_id'], replace=True, size=n_negative
)
other_features = np.random.choice(df['age'], replace=True, size=n_negative)

# Construct negative pairs DataFrame
negative_pairs = pd.DataFrame(
    {'article_id': article_ids, 'customer_id': customer_ids, 'age': other_features}
)

Next, we add the labels and concatenate (vertically) the positive and negative pairs into one DataFrame

positive_pairs['label'] = 1
negative_pairs['label'] = 0

ranking_df = pd.concat([positive_pairs, negative_pairs])

ranking_df.head(5)

### Create Item Features

item_features = [
    'article_id',
    'product_type_name',
    'product_group_name',
    'graphical_appearance_name',
    'colour_group_name',
    'perceived_colour_value_name',
    'perceived_colour_master_name',
    'department_name',
    'index_name',
    'index_group_name',
    'section_name',
    'garment_group_name',
]

item_df = all_articles_df.copy()
item_df = item_df[item_features]

### Create Ranking Dataset

ranking_df = ranking_df.merge(item_df, on='article_id', how='left')

ranking_df.head(5)

ranking_df.isnull().sum()

### Save the Engineered Features
Lastly, we pickle it to save the generated ranking dataset

ranking_df.to_pickle(processed_data_path / 'ranking.pkl')
